requirements.txt
ultralytics
flask
Pillow